+++ apps/modernjs/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @module-federation/modernjs
 
+## 0.1.13
+
+### Patch Changes
+
+- Updated dependencies [56eb56a]
+- Updated dependencies [b120511]
+  - @module-federation/enhanced@0.1.8
+
+## 0.1.12
+
+### Patch Changes
+
+- Updated dependencies [648353b]
+  - @module-federation/enhanced@0.1.7
+
 ## 0.1.11
 
 ### Patch Changes