+++ packages/dts-plugin/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @module-federation/dts-plugin
 
+## 0.1.8
+
+### Patch Changes
+
+- @module-federation/sdk@0.1.8
+- @module-federation/managers@0.1.8
+- @module-federation/third-party-dts-extractor@0.1.8
+
+## 0.1.7
+
+### Patch Changes
+
+- @module-federation/sdk@0.1.7
+- @module-federation/managers@0.1.7
+- @module-federation/third-party-dts-extractor@0.1.7
+
 ## 0.1.6
 
 ### Patch Changes